requirements.txt
celery==5.2.7
Django==4.1.5
django-debug-toolbar==3.8.1
django-environ==0.9.0
django-filter==22.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
django-redis==5.2.0
drf-yasg==1.21.4
flower==1.2.0
furo==2022.12.7
psycopg2-binary==2.9.5
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
requests==2.28.2
sentry-sdk==1.14.0
Sphinx==6.1.3